fix(audit-fix): RES-03 Payable Function May Lead To Minting Of Less Tokens
The function deposit() is marked as payable, allowing for transfer of
native Ether into the smart contract. This function further calls the
internal function _deposit() while providing _amountIn as an input
parameter. This effectively means that the amount of minted tokens for
the user will solely depend on the provided input parameter _amountIn.
If the user mistakenly sends Ether when calling this function, they
will not be able to mint those funds as tokens.